conflicts: materialize conflict labels in working copy
I swapped the order of added and removed term in the label for diff
sections, because I think having the added term always come first makes
conflicts easier to read.

For instance, if I see this conflict:

```
<<<<<<< Conflict 1 of 1
%%%%%%% Changes from BBBBBBBB to AAAAAAAA
...
+++++++ Contents of CCCCCCCC
...
>>>>>>> Conflict 1 of 1 ends
```

It's less clear to me that the two added sides are "AAAAAAAA" and
"CCCCCCCC" compared to this wording:

```
<<<<<<< conflict 1 of 1
%%%%%%% AAAAAAAA compared with BBBBBBBB
...
+++++++ CCCCCCCC
...
>>>>>>> conflict 1 of 1 ends
```

This will become more important in later commits, which replace the
"side" and "base" labels with change/commit IDs.